proposed changes to README.md
I tried to make the read.me a bit less confusing. We were separating two things that in fact are the same thing: (i) people looking to get funding for their project and (ii) people who have an idea that could benefit the ecosystem. Everything has to benefit Aragon or Ethereum and if it is just an idea or a project people are already working on, the process is the same. The thing that would change is the proposal itself. Also, we were explaining twice the process to follow for when people were working on something that already had an existing proposal.